anchor messages: relay, rpc and validation. WIP

+++ src/spv/support/BRInt.h
@@ -60,22 +60,22 @@ extern "C" {
     _hexc((u).u8[30] >> 4), _hexc((u).u8[30]), _hexc((u).u8[31] >> 4), _hexc((u).u8[31]), '\0' })
 
 #define toUInt256(s) ((const UInt256) { .u8 = {\
-    (_hexu((s)[ 0]) << 4) | _hexu((s)[ 1]), (_hexu((s)[ 2]) << 4) | _hexu((s)[ 3]),\
-    (_hexu((s)[ 4]) << 4) | _hexu((s)[ 5]), (_hexu((s)[ 6]) << 4) | _hexu((s)[ 7]),\
-    (_hexu((s)[ 8]) << 4) | _hexu((s)[ 9]), (_hexu((s)[10]) << 4) | _hexu((s)[11]),\
-    (_hexu((s)[12]) << 4) | _hexu((s)[13]), (_hexu((s)[14]) << 4) | _hexu((s)[15]),\
-    (_hexu((s)[16]) << 4) | _hexu((s)[17]), (_hexu((s)[18]) << 4) | _hexu((s)[19]),\
-    (_hexu((s)[20]) << 4) | _hexu((s)[21]), (_hexu((s)[22]) << 4) | _hexu((s)[23]),\
-    (_hexu((s)[24]) << 4) | _hexu((s)[25]), (_hexu((s)[26]) << 4) | _hexu((s)[27]),\
-    (_hexu((s)[28]) << 4) | _hexu((s)[29]), (_hexu((s)[30]) << 4) | _hexu((s)[31]),\
-    (_hexu((s)[32]) << 4) | _hexu((s)[33]), (_hexu((s)[34]) << 4) | _hexu((s)[35]),\
-    (_hexu((s)[36]) << 4) | _hexu((s)[37]), (_hexu((s)[38]) << 4) | _hexu((s)[39]),\
-    (_hexu((s)[40]) << 4) | _hexu((s)[41]), (_hexu((s)[42]) << 4) | _hexu((s)[43]),\
-    (_hexu((s)[44]) << 4) | _hexu((s)[45]), (_hexu((s)[46]) << 4) | _hexu((s)[47]),\
-    (_hexu((s)[48]) << 4) | _hexu((s)[49]), (_hexu((s)[50]) << 4) | _hexu((s)[51]),\
-    (_hexu((s)[52]) << 4) | _hexu((s)[53]), (_hexu((s)[54]) << 4) | _hexu((s)[55]),\
-    (_hexu((s)[56]) << 4) | _hexu((s)[57]), (_hexu((s)[58]) << 4) | _hexu((s)[59]),\
-    (_hexu((s)[60]) << 4) | _hexu((s)[61]), (_hexu((s)[62]) << 4) | _hexu((s)[63]) } })
+    (uint8_t)((_hexu((s)[ 0]) << 4) | _hexu((s)[ 1])), (uint8_t)((_hexu((s)[ 2]) << 4) | _hexu((s)[ 3])),\
+    (uint8_t)((_hexu((s)[ 4]) << 4) | _hexu((s)[ 5])), (uint8_t)((_hexu((s)[ 6]) << 4) | _hexu((s)[ 7])),\
+    (uint8_t)((_hexu((s)[ 8]) << 4) | _hexu((s)[ 9])), (uint8_t)((_hexu((s)[10]) << 4) | _hexu((s)[11])),\
+    (uint8_t)((_hexu((s)[12]) << 4) | _hexu((s)[13])), (uint8_t)((_hexu((s)[14]) << 4) | _hexu((s)[15])),\
+    (uint8_t)((_hexu((s)[16]) << 4) | _hexu((s)[17])), (uint8_t)((_hexu((s)[18]) << 4) | _hexu((s)[19])),\
+    (uint8_t)((_hexu((s)[20]) << 4) | _hexu((s)[21])), (uint8_t)((_hexu((s)[22]) << 4) | _hexu((s)[23])),\
+    (uint8_t)((_hexu((s)[24]) << 4) | _hexu((s)[25])), (uint8_t)((_hexu((s)[26]) << 4) | _hexu((s)[27])),\
+    (uint8_t)((_hexu((s)[28]) << 4) | _hexu((s)[29])), (uint8_t)((_hexu((s)[30]) << 4) | _hexu((s)[31])),\
+    (uint8_t)((_hexu((s)[32]) << 4) | _hexu((s)[33])), (uint8_t)((_hexu((s)[34]) << 4) | _hexu((s)[35])),\
+    (uint8_t)((_hexu((s)[36]) << 4) | _hexu((s)[37])), (uint8_t)((_hexu((s)[38]) << 4) | _hexu((s)[39])),\
+    (uint8_t)((_hexu((s)[40]) << 4) | _hexu((s)[41])), (uint8_t)((_hexu((s)[42]) << 4) | _hexu((s)[43])),\
+    (uint8_t)((_hexu((s)[44]) << 4) | _hexu((s)[45])), (uint8_t)((_hexu((s)[46]) << 4) | _hexu((s)[47])),\
+    (uint8_t)((_hexu((s)[48]) << 4) | _hexu((s)[49])), (uint8_t)((_hexu((s)[50]) << 4) | _hexu((s)[51])),\
+    (uint8_t)((_hexu((s)[52]) << 4) | _hexu((s)[53])), (uint8_t)((_hexu((s)[54]) << 4) | _hexu((s)[55])),\
+    (uint8_t)((_hexu((s)[56]) << 4) | _hexu((s)[57])), (uint8_t)((_hexu((s)[58]) << 4) | _hexu((s)[59])),\
+    (uint8_t)((_hexu((s)[60]) << 4) | _hexu((s)[61])), (uint8_t)((_hexu((s)[62]) << 4) | _hexu((s)[63])) } })
 
 #define _hexc(u) ((char) (((u) & 0x0f) + ((((u) & 0x0f) <= 9) ? '0' : 'a' - 0x0a)))
 #define _hexu(c) (((c) >= '0' && (c) <= '9') ? (c) - '0' : ((c) >= 'a' && (c) <= 'f') ? (c) - ('a' - 0x0a) :\
